Move task subclasses into their own files
