+++ app/models/tasks/aod_motion_mail_task.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+class AodMotionMailTask < MailTask
+  def self.label
+    COPY::AOD_MOTION_MAIL_TASK_LABEL
+  end
+
+  def self.default_assignee(_parent, _params)
+    AodTeam.singleton
+  end
+end
