+++ app/models/tasks/appeal_withdrawal_mail_task.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+class AppealWithdrawalMailTask < MailTask
+  def self.label
+    COPY::APPEAL_WITHDRAWAL_MAIL_TASK_LABEL
+  end
+
+  def self.default_assignee(_parent, _params)
+    Colocated.singleton
+  end
+end
