+++ app/models/tasks/controlled_correspondence_mail_task.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+class ControlledCorrespondenceMailTask < MailTask
+  def self.label
+    COPY::CONTROLLED_CORRESPONDENCE_MAIL_TASK_LABEL
+  end
+
+  def self.default_assignee(_parent, _params)
+    LitigationSupport.singleton
+  end
+end
