+++ app/models/tasks/death_certificate_mail_task.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+class DeathCertificateMailTask < MailTask
+  def self.label
+    COPY::DEATH_CERTIFICATE_MAIL_TASK_LABEL
+  end
+
+  def self.default_assignee(_parent, _params)
+    Colocated.singleton
+  end
+end
