+++ app/models/tasks/hearing_admin_action_contested_claimant_task.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class HearingAdminActionContestedClaimantTask < HearingAdminActionTask
+  def self.label
+    "Contested claimant issue"
+  end
+end
