+++ app/models/tasks/hearing_admin_action_foia_privacy_request_task.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class HearingAdminActionFoiaPrivacyRequestTask < HearingAdminActionTask
+  def self.label
+    "FOIA/Privacy request"
+  end
+end
