+++ app/models/tasks/hearing_admin_action_incarcerated_veteran_task.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class HearingAdminActionIncarceratedVeteranTask < HearingAdminActionTask
+  def self.label
+    "Veteran is incarcerated"
+  end
+end
