+++ app/models/tasks/hearing_admin_action_missing_forms_task.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class HearingAdminActionMissingFormsTask < HearingAdminActionTask
+  def self.label
+    "Missing forms"
+  end
+end
