+++ app/models/tasks/hearing_admin_action_other_task.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class HearingAdminActionOtherTask < HearingAdminActionTask
+  def self.label
+    "Other"
+  end
+end
