+++ app/models/tasks/hearing_admin_action_verify_poa_task.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class HearingAdminActionVerifyPoaTask < HearingAdminActionTask
+  def self.label
+    "Verify power of attorney"
+  end
+end
