+++ app/models/tasks/hearing_related_mail_task.rb
@@ -1,0 +1,19 @@
+# frozen_string_literal: true
+
+class HearingRelatedMailTask < MailTask
+  def self.blocking?
+    true
+  end
+
+  def self.label
+    COPY::HEARING_RELATED_MAIL_TASK_LABEL
+  end
+
+  def self.default_assignee(parent, _params)
+    fail Caseflow::Error::MailRoutingError unless case_active?(parent)
+
+    return HearingAdmin.singleton if pending_hearing_task?(parent)
+
+    Colocated.singleton
+  end
+end
