+++ app/models/tasks/mail_task.rb
@@ -86,267 +86,3 @@ class MailTask < GenericTask
   #   super(user).unshift(Constants.TASK_ACTIONS.CHANGE_TASK_TYPE.to_h)
   # end
 end
-
-class AddressChangeMailTask < MailTask
-  def self.label
-    COPY::ADDRESS_CHANGE_MAIL_TASK_LABEL
-  end
-
-  def self.default_assignee(parent, _params)
-    fail Caseflow::Error::MailRoutingError unless case_active?(parent)
-
-    return HearingAdmin.singleton if pending_hearing_task?(parent)
-
-    Colocated.singleton
-  end
-end
-
-class AodMotionMailTask < MailTask
-  def self.label
-    COPY::AOD_MOTION_MAIL_TASK_LABEL
-  end
-
-  def self.default_assignee(_parent, _params)
-    AodTeam.singleton
-  end
-end
-
-class AppealWithdrawalMailTask < MailTask
-  def self.label
-    COPY::APPEAL_WITHDRAWAL_MAIL_TASK_LABEL
-  end
-
-  def self.default_assignee(_parent, _params)
-    Colocated.singleton
-  end
-end
-
-class ClearAndUnmistakeableErrorMailTask < MailTask
-  def available_actions(user)
-    if LitigationSupport.singleton.user_has_access?(user)
-      return super.push(Constants.TASK_ACTIONS.LIT_SUPPORT_PULAC_CERULLO.to_h)
-    end
-
-    super
-  end
-
-  def self.label
-    COPY::CLEAR_AND_UNMISTAKABLE_ERROR_MAIL_TASK_LABEL
-  end
-
-  def self.default_assignee(_parent, _params)
-    LitigationSupport.singleton
-  end
-end
-
-class CongressionalInterestMailTask < MailTask
-  def self.blocking?
-    true
-  end
-
-  def self.label
-    COPY::CONGRESSIONAL_INTEREST_MAIL_TASK_LABEL
-  end
-
-  def self.default_assignee(_parent, _params)
-    LitigationSupport.singleton
-  end
-end
-
-class ControlledCorrespondenceMailTask < MailTask
-  def self.label
-    COPY::CONTROLLED_CORRESPONDENCE_MAIL_TASK_LABEL
-  end
-
-  def self.default_assignee(_parent, _params)
-    LitigationSupport.singleton
-  end
-end
-
-class DeathCertificateMailTask < MailTask
-  def self.label
-    COPY::DEATH_CERTIFICATE_MAIL_TASK_LABEL
-  end
-
-  def self.default_assignee(_parent, _params)
-    Colocated.singleton
-  end
-end
-
-class EvidenceOrArgumentMailTask < MailTask
-  def self.label
-    COPY::EVIDENCE_OR_ARGUMENT_MAIL_TASK_LABEL
-  end
-
-  def self.default_assignee(parent, _params)
-    fail Caseflow::Error::MailRoutingError unless case_active?(parent)
-
-    return HearingAdmin.singleton if pending_hearing_task?(parent)
-
-    Colocated.singleton
-  end
-end
-
-class ExtensionRequestMailTask < MailTask
-  def self.blocking?
-    true
-  end
-
-  def self.label
-    COPY::EXTENSION_REQUEST_MAIL_TASK_LABEL
-  end
-
-  def self.default_assignee(parent, _params)
-    fail Caseflow::Error::MailRoutingError unless case_active?(parent)
-
-    Colocated.singleton
-  end
-end
-
-class FoiaRequestMailTask < MailTask
-  def self.blocking?
-    true
-  end
-
-  def self.label
-    COPY::FOIA_REQUEST_MAIL_TASK_LABEL
-  end
-
-  def self.default_assignee(_parent, _params)
-    PrivacyTeam.singleton
-  end
-end
-
-class HearingRelatedMailTask < MailTask
-  def self.blocking?
-    true
-  end
-
-  def self.label
-    COPY::HEARING_RELATED_MAIL_TASK_LABEL
-  end
-
-  def self.default_assignee(parent, _params)
-    fail Caseflow::Error::MailRoutingError unless case_active?(parent)
-
-    return HearingAdmin.singleton if pending_hearing_task?(parent)
-
-    Colocated.singleton
-  end
-end
-
-class OtherMotionMailTask < MailTask
-  def self.label
-    COPY::OTHER_MOTION_MAIL_TASK_LABEL
-  end
-
-  def self.default_assignee(_parent, _params)
-    LitigationSupport.singleton
-  end
-end
-
-class PowerOfAttorneyRelatedMailTask < MailTask
-  def self.blocking?
-    true
-  end
-
-  def self.label
-    COPY::POWER_OF_ATTORNEY_MAIL_TASK_LABEL
-  end
-
-  def self.default_assignee(parent, _params)
-    fail Caseflow::Error::MailRoutingError unless case_active?(parent)
-
-    return HearingAdmin.singleton if pending_hearing_task?(parent)
-
-    Colocated.singleton
-  end
-end
-
-class PrivacyActRequestMailTask < MailTask
-  def self.blocking?
-    true
-  end
-
-  def self.label
-    COPY::PRIVACY_ACT_REQUEST_MAIL_TASK_LABEL
-  end
-
-  def self.default_assignee(_parent, _params)
-    PrivacyTeam.singleton
-  end
-end
-
-class PrivacyComplaintMailTask < MailTask
-  def self.blocking?
-    true
-  end
-
-  def self.label
-    COPY::PRIVACY_COMPLAINT_MAIL_TASK_LABEL
-  end
-
-  def self.default_assignee(_parent, _params)
-    PrivacyTeam.singleton
-  end
-end
-
-class ReturnedUndeliverableCorrespondenceMailTask < MailTask
-  def self.label
-    COPY::RETURNED_CORRESPONDENCE_MAIL_TASK_LABEL
-  end
-
-  def self.default_assignee(parent, _params)
-    return BvaDispatch.singleton if !case_active?(parent)
-    return HearingAdmin.singleton if pending_hearing_task?(parent)
-    return most_recent_active_task_assignee(parent) if most_recent_active_task_assignee(parent)
-
-    fail Caseflow::Error::MailRoutingError
-  end
-end
-
-class ReconsiderationMotionMailTask < MailTask
-  def available_actions(user)
-    if LitigationSupport.singleton.user_has_access?(user)
-      return super.push(Constants.TASK_ACTIONS.LIT_SUPPORT_PULAC_CERULLO.to_h)
-    end
-
-    super
-  end
-
-  def self.label
-    COPY::RECONSIDERATION_MOTION_MAIL_TASK_LABEL
-  end
-
-  def self.default_assignee(_parent, _params)
-    LitigationSupport.singleton
-  end
-end
-
-class StatusInquiryMailTask < MailTask
-  def self.label
-    COPY::STATUS_INQUIRY_MAIL_TASK_LABEL
-  end
-
-  def self.default_assignee(_parent, _params)
-    LitigationSupport.singleton
-  end
-end
-
-class VacateMotionMailTask < MailTask
-  def available_actions(user)
-    if LitigationSupport.singleton.user_has_access?(user)
-      return super.push(Constants.TASK_ACTIONS.LIT_SUPPORT_PULAC_CERULLO.to_h)
-    end
-
-    super
-  end
-
-  def self.label
-    COPY::VACATE_MOTION_MAIL_TASK_LABEL
-  end
-
-  def self.default_assignee(_parent, _params)
-    LitigationSupport.singleton
-  end
-end
