+++ app/models/tasks/other_motion_mail_task.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+class OtherMotionMailTask < MailTask
+  def self.label
+    COPY::OTHER_MOTION_MAIL_TASK_LABEL
+  end
+
+  def self.default_assignee(_parent, _params)
+    LitigationSupport.singleton
+  end
+end
