+++ app/models/tasks/power_of_attorney_related_mail_task.rb
@@ -1,0 +1,19 @@
+# frozen_string_literal: true
+
+class PowerOfAttorneyRelatedMailTask < MailTask
+  def self.blocking?
+    true
+  end
+
+  def self.label
+    COPY::POWER_OF_ATTORNEY_MAIL_TASK_LABEL
+  end
+
+  def self.default_assignee(parent, _params)
+    fail Caseflow::Error::MailRoutingError unless case_active?(parent)
+
+    return HearingAdmin.singleton if pending_hearing_task?(parent)
+
+    Colocated.singleton
+  end
+end
