+++ app/models/tasks/privacy_act_request_mail_task.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+class PrivacyActRequestMailTask < MailTask
+  def self.blocking?
+    true
+  end
+
+  def self.label
+    COPY::PRIVACY_ACT_REQUEST_MAIL_TASK_LABEL
+  end
+
+  def self.default_assignee(_parent, _params)
+    PrivacyTeam.singleton
+  end
+end
