+++ app/models/tasks/status_inquiry_mail_task.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+class StatusInquiryMailTask < MailTask
+  def self.label
+    COPY::STATUS_INQUIRY_MAIL_TASK_LABEL
+  end
+
+  def self.default_assignee(_parent, _params)
+    LitigationSupport.singleton
+  end
+end
