+++ app/models/tasks/vacate_motion_mail_task.rb
@@ -1,0 +1,19 @@
+# frozen_string_literal: true
+
+class VacateMotionMailTask < MailTask
+  def available_actions(user)
+    if LitigationSupport.singleton.user_has_access?(user)
+      return super.push(Constants.TASK_ACTIONS.LIT_SUPPORT_PULAC_CERULLO.to_h)
+    end
+
+    super
+  end
+
+  def self.label
+    COPY::VACATE_MOTION_MAIL_TASK_LABEL
+  end
+
+  def self.default_assignee(_parent, _params)
+    LitigationSupport.singleton
+  end
+end
